MBRL/requirements.txt
absl-py==2.1.0
chex==0.1.87
jax>=0.4.35
jaxlib==0.4.35
ml-dtypes==0.5.0
numpy==2.1.3
opt-einsum==3.4.0
pygame==2.6.1
scipy==1.14.1
toolz==1.0.0
typing-extensions==4.12.2
gymnax==0.0.8
dm-haiku
tqdm
